[DB migrations] bugfixes related to setup from scratch

+++ db/migrate/20171006135205_cleanup_old_table.rb
@@ -1,6 +1,14 @@
-class CleanupOldTable < ActiveRecord::Migration
+class CleanupOldTable < ActiveRecord::Migration[5.1]
   def up
-    drop_table "_event_logs_old"
+    sql = <<-eos
+      SELECT EXISTS(
+        SELECT * from information_schema.tables where table_name='_event_logs_old'
+      )
+    eos
+
+    if ActiveRecord::Base.connection.execute(sql.strip).first["exists"] == "t"
+      drop_table "_event_logs_old"
+    end
   end
 
   def down
